backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-rest-knox
django-rest-passwordreset
Pillow
python-dotenv
google-genai